day5/pyproject.toml
#create toml file

[project]
name = "banking_project"
version = "0.0.1"
description = "A simple banking project implemented in Python"
authors = [
    { name = "Salmaan", email = "salman@abc.com"}
]
readme = "README.md"
requires-python = ">3.8"

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "faker>=18.0.0",
    "pandas>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "black>=23.3.0",
    "isort>=5.10.1",
]

[tool.black]
line = 88
target-version = ['py313']